requirements.txt
Flask==1.1.2
Flask-Bootstrap4==4.0.2
Flask-WTF==0.14.3
openai==0.2.4
twilio==6.44.2
WTForms==2.3.3
python-dotenv==0.17.1
